# Conflicts:
#	drivers/counter/counter_ll_stm32_rtc.c
#	drivers/ethernet/eth_stm32_hal.c
#	drivers/gpio/gpio_shell.c
#	drivers/i2c/i2c_shell.c
#	drivers/modem/modem_cellular.c
#	drivers/pinctrl/pinctrl_stm32.c
#	drivers/rtc/rtc_ll_stm32.c
#	drivers/sensor/grow_r502a/grow_r502a.h
#	dts/arm/st/h7/stm32h7.dtsi
#	dts/bindings/mdio/st,stm32-mdio.yaml
#	include/zephyr/net/net_context.h
#	include/zephyr/net/net_if.h
#	include/zephyr/smf.h
#	kernel/sem.c
#	subsys/net/ip/ipv6.h
#	subsys/net/ip/ipv6_mld.c
#	subsys/net/ip/ipv6_nbr.c
#	subsys/net/ip/net_context.c
#	subsys/net/ip/net_if.c
#	subsys/net/ip/tcp.c
#	subsys/net/ip/tcp_private.h
#	subsys/net/lib/shell/ipv4.c
#	subsys/net/lib/sockets/sockets.c
#	subsys/shell/backends/shell_uart.c

+++ subsys/net/buf_simple.c
@@ -34,8 +34,8 @@ void net_buf_simple_init_with_data(struct net_buf_simple *buf,
 {
 	buf->__buf = data;
 	buf->data  = data;
-	buf->size  = size;
-	buf->len   = size;
+	buf->size  = (uint16_t)size;
+	buf->len   = (uint16_t)size;
 }
 
 void net_buf_simple_reserve(struct net_buf_simple *buf, size_t reserve)
@@ -61,7 +61,7 @@ void *net_buf_simple_add(struct net_buf_simple *buf, size_t len)
 
 	__ASSERT_NO_MSG(net_buf_simple_tailroom(buf) >= len);
 
-	buf->len += len;
+	buf->len += (uint16_t)len;
 	return tail;
 }
 
@@ -175,7 +175,7 @@ void *net_buf_simple_remove_mem(struct net_buf_simple *buf, size_t len)
 
 	__ASSERT_NO_MSG(buf->len >= len);
 
-	buf->len -= len;
+	buf->len -= (uint16_t)len;
 	return buf->data + buf->len;
 }
 
@@ -341,7 +341,7 @@ void *net_buf_simple_push(struct net_buf_simple *buf, size_t len)
 	__ASSERT_NO_MSG(net_buf_simple_headroom(buf) >= len);
 
 	buf->data -= len;
-	buf->len += len;
+	buf->len += (uint16_t)len;
 	return buf->data;
 }
 
@@ -450,7 +450,7 @@ void *net_buf_simple_pull(struct net_buf_simple *buf, size_t len)
 
 	__ASSERT_NO_MSG(buf->len >= len);
 
-	buf->len -= len;
+	buf->len -= (uint16_t)len;
 	return buf->data += len;
 }
 
@@ -462,7 +462,7 @@ void *net_buf_simple_pull_mem(struct net_buf_simple *buf, size_t len)
 
 	__ASSERT_NO_MSG(buf->len >= len);
 
-	buf->len -= len;
+	buf->len -= (uint16_t)len;
 	buf->data += len;
 
 	return data;
@@ -622,5 +622,5 @@ size_t net_buf_simple_tailroom(struct net_buf_simple *buf)
 
 uint16_t net_buf_simple_max_len(struct net_buf_simple *buf)
 {
-	return buf->size - net_buf_simple_headroom(buf);
+	return buf->size - (uint16_t)net_buf_simple_headroom(buf);
 }
